+++ plugins/src/main/scala/com/stratio/sparta/plugin/workflow/input/avro/AvroInputStepBatch.scala
@@ -46,6 +46,12 @@ class AvroInputStepBatch(
         messages = validation.messages :+ WorkflowValidationMessage(s"the input path cannot be empty", name)
       )
 
+    if(debugOptions.isDefined && !validDebuggingOptions)
+      validation = ErrorValidations(
+        valid = false,
+        messages = validation.messages :+ WorkflowValidationMessage(s"$errorDebugValidation", name)
+      )
+
     validation
   }