+++ plugins/src/main/scala/com/stratio/sparta/plugin/workflow/input/csv/CsvInputStepBatch.scala
@@ -40,6 +40,12 @@ class CsvInputStepBatch(
         messages = validation.messages :+ WorkflowValidationMessage(s"the input path cannot be empty", name)
       )
 
+    if(debugOptions.isDefined && !validDebuggingOptions)
+      validation = ErrorValidations(
+        valid = false,
+        messages = validation.messages :+ WorkflowValidationMessage(s"$errorDebugValidation", name)
+      )
+
     validation
   }