+++ plugins/src/main/scala/com/stratio/sparta/plugin/workflow/input/filesystem/FileSystemInputStepStreaming.scala
@@ -47,6 +47,12 @@ class FileSystemInputStepStreaming(
         messages = validation.messages :+ WorkflowValidationMessage(s"the path cannot be empty", name)
       )
 
+    if(debugOptions.isDefined && !validDebuggingOptions)
+      validation = ErrorValidations(
+        valid = false,
+        messages = validation.messages :+ WorkflowValidationMessage(s"$errorDebugValidation", name)
+      )
+
     validation
   }