+++ plugins/src/main/scala/com/stratio/sparta/plugin/workflow/input/kafka/KafkaInputStepStreaming.scala
@@ -93,6 +93,12 @@ class KafkaInputStepStreaming(
           WorkflowValidationMessage(s"the $REQUEST_TIMEOUT_MS_CONFIG should be greater than $FETCH_MAX_WAIT_MS_CONFIG", name)
       )
 
+    if(debugOptions.isDefined && !validDebuggingOptions)
+      validation = ErrorValidations(
+        valid = false,
+        messages = validation.messages :+ WorkflowValidationMessage(s"$errorDebugValidation", name)
+      )
+
     validation
   }