+++ plugins/src/main/scala/com/stratio/sparta/plugin/workflow/input/websocket/WebSocketInputStepStreaming.scala
@@ -43,6 +43,12 @@ class WebSocketInputStepStreaming(
         messages = validation.messages :+ WorkflowValidationMessage(s"url cannot be empty", name)
       )
 
+    if(debugOptions.isDefined && !validDebuggingOptions)
+      validation = ErrorValidations(
+        valid = false,
+        messages = validation.messages :+ WorkflowValidationMessage(s"$errorDebugValidation", name)
+      )
+
     validation
   }