+++ sdk/src/main/scala/com/stratio/sparta/sdk/workflow/step/DebugOptions.scala
@@ -3,11 +3,8 @@
  *
  * This software – including all its source code – contains proprietary information of Stratio Big Data Inc., Sucursal en España and may not be revealed, sold, transferred, modified, distributed or otherwise made available, licensed or sublicensed to third parties; nor reverse engineered, disassembled or decompiled, without express written authorization from Stratio Big Data Inc., Sucursal en España.
  */
-package com.stratio.sparta.sdk.enumerators
+package com.stratio.sparta.sdk.workflow.step
 
-object InputFormatEnum extends Enumeration {
-
-  type Status = Value
-  val JSON, STRING, AVRO, CSV, PARQUET = Value
-
-}
+case class DebugOptions( path: Option[String] = None,
+                         query: Option[String] = None,
+                         userProvidedExample: Option[String] = None)